+++ inc/PID.h
@@ -13,11 +13,11 @@
 #include "Interrupt.h"
 
 typedef struct {
-	float kp;
-	float ki;
-	float kd; //PID values
+	int32_t kp;
+	int32_t ki;
+	int32_t kd; //PID values
 } pidSettingsType;
 
-extern float computePID(uint16_t const currentValue, uint16_t setpoint);
+extern int32_t computePID(uint16_t const currentValue, uint16_t setpoint);
 extern void setupPID(void);
 #endif /* PID_H_ */
